Add files via upload
This is Version 1.0 of an ontoterminology (i.e. 'a terminology whose conceptual system is a formal ontology' Roche 2007 ) that defines and represents the archaeological finds of T-Pillars in Göbekli Tepe, SE Turkey (-10000 to -8300 BCE) , in a machine-tractable way.
This dataset was created by Antonia Lourentzaki (Department of Archaeology, University of Crete), under the supervision of Dr. Maria Papadopoulou (Department of Philology, University of Crete). 
We gratefully acknowledge the generous funding of the European Commission under the ERA Chair Project TALOS AI for SSH (Grant agreement ID: 101087269).