Add files via upload
This is version 1.0 of the ALyrA ontoterminology (i.e., 'a terminology whose conceptual system is a formal ontology', Roche 2007 ), which focuses on modelling Archaic Lyric poets (799-430 BC). AlyrA stands for "Archaic Lyrical Agora". 
The ontoterminology was created by Rafail Giannadakis (Department of Philology, University of Crete) under the supervision of Dr. Maria Papadopoulou (Department of Philology, University of Crete).
We gratefully acknowledge the generous funding of the European Commission in the context of the TALOS AI for SSH project (Grant agreement ID: 101087269).